requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-flask>=1.2.0
requests>=2.31.0

# Code formatting
black>=22.0.0
isort>=5.10.0

# Linting
flake8>=4.0.0
pylint>=2.12.0

# Type checking
mypy>=0.950

# Environment management
python-dotenv>=0.19.0
